Update main diary page to populate entries from mock backend JSON.
The 'BACKEND' constant helps to retrieve the mock backend JSON data from the environment variables (on mobile dev app via USB) or directly fetch from the local host (on the PC browser view). A simple .env file is required for this as well as starting the JSON server using the command 'npx json-server assets/db.json'.